Removes OpenFEC and MockOpenNFEC apps, rolling their functionality into the committee_accounts and committee apps
