+++ django-backend/fecfiler/contacts/tests/test_views.py
@@ -311,51 +311,52 @@ class ContactViewSetTest(TestCase):
         self.assertEqual(response.status_code, 400)
 
     def test_get_committee_mocked(self):
-        with patch("fecfiler.contacts.views.mock_committee") as mock_committee:
-            mock_committee.return_value = {"name": "TEST"}
+        with patch("fecfiler.committee_accounts.utils.settings") as settings:
+            settings.FLAG__COMMITTEE_DATA_SOURCE = "REDIS"
+            with patch(
+                "fecfiler.committee_accounts.utils.get_committee_from_redis"
+            ) as mock_committee:
+                mock_committee.return_value = {"name": "TEST"}
 
-            request = self.factory.get(
-                "/api/v1/contacts/committee/?committee_id=C12345678"
-            )
-            request.user = self.user
-            request.session = {
-                "committee_uuid": "11111111-2222-3333-4444-555555555555",
-                "committee_id": "C01234567",
-            }
+                request = self.factory.get(
+                    "/api/v1/contacts/committee/?committee_id=C12345678"
+                )
+                request.user = self.user
+                request.session = {
+                    "committee_uuid": "11111111-2222-3333-4444-555555555555",
+                    "committee_id": "C01234567",
+                }
 
-            response = ContactViewSet.as_view({"get": "committee"})(request)
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.data["name"], "TEST")
+                response = ContactViewSet.as_view({"get": "committee"})(request)
+                self.assertEqual(response.status_code, 200)
+                self.assertEqual(response.data["name"], "TEST")
 
     def test_get_committee_from_efo(self):
-        with patch("fecfiler.contacts.views.settings") as settings:
+        with patch("fecfiler.committee_accounts.utils.settings") as settings:
+            settings.FLAG__COMMITTEE_DATA_SOURCE = "PRODUCTION"
             settings.FEC_API = "https://not-real.api/"
             settings.FEC_API_KEY = "FAKE-KEY"
-
             expected_call = 'https://not-real.api/committee/C12345678/?api_key=FAKE-KEY'
-            with patch("fecfiler.contacts.views.mock_committee") as mock_committee:
-                mock_committee.return_value = None
+            with patch("fecfiler.committee_accounts.utils.requests") as mock_requests:
+                mock_requests.get = Mock()
+                mock_response = Mock()
+                mock_response.json = Mock()
+                mock_response.json.return_value = {"name": "TEST"}
+                mock_requests.get.return_value = mock_response
 
-                with patch("fecfiler.contacts.views.requests") as mock_requests:
-                    mock_requests.get = Mock()
-                    mock_response = Mock()
-                    mock_response.json = Mock()
-                    mock_response.json.return_value = {"name": "TEST"}
-                    mock_requests.get.return_value = mock_response
+                request = self.factory.get(
+                    "/api/v1/contacts/committee/?committee_id=C12345678"
+                )
+                request.user = self.user
+                request.session = {
+                    "committee_uuid": "11111111-2222-3333-4444-555555555555",
+                    "committee_id": "C01234567",
+                }
 
-                    request = self.factory.get(
-                        "/api/v1/contacts/committee/?committee_id=C12345678"
-                    )
-                    request.user = self.user
-                    request.session = {
-                        "committee_uuid": "11111111-2222-3333-4444-555555555555",
-                        "committee_id": "C01234567",
-                    }
-
-                    response = ContactViewSet.as_view({"get": "committee"})(request)
-                    self.assertEqual(response.status_code, 200)
-                    self.assertIn(expected_call, mock_requests.get.call_args[0])
-                    self.assertEqual(response.data["name"], "TEST")
+                response = ContactViewSet.as_view({"get": "committee"})(request)
+                self.assertEqual(response.status_code, 200)
+                self.assertIn(expected_call, mock_requests.get.call_args[0])
+                self.assertEqual(response.data["name"], "TEST")
 
     def test_restore_no_match(self):
         request = self.factory.post(
