--- django-backend/fecfiler/mock_openfec/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MockOpenfecConfig(AppConfig):
-    name = "fecfiler.mock_openfec"
