--- django-backend/fecfiler/openfec/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class OpenfecConfig(AppConfig):
-    name = 'fecfiler.openfec'
