+++ django-backend/fecfiler/settings/base.py
@@ -76,9 +76,7 @@ INSTALLED_APPS = [
     "fecfiler.soft_delete",
     "fecfiler.validation",
     "fecfiler.web_services",
-    "fecfiler.openfec",
     "fecfiler.user",
-    "fecfiler.mock_openfec",
     "fecfiler.oidc",
     "fecfiler.devops",
     "fecfiler.mock_oidc_provider",
